# Faster fully depthwise-separable 1D conv