Benchmarks:


Kernel Size | Input Steps | Pre (sec | Post (sec)
--- | ---- | ---- | ---- |
3 | 1 |0.152 | 0.058
11 | 1 | 0.144 | 0.071
3 | 1024 |  3.349 | 0.228
11 | 1024 | 3.347 | 0.394

closes https://github.com/ml-explore/mlx/issues/2369